change formula to nws flow and use #define for coeff

+++ src/EnvironmentCalculations.h
@@ -70,13 +70,11 @@ namespace EnvironmentCalculations
   /////////////////////////////////////////////////////////////////
   /// Calculate the heatindex based on the humidity and temperature
   /// in tempUnit.
-  /// the heatindex does work for values above 26.7°C/80°F and 40% humidity.
-  /// For values below  26.7°C/80°F use windchill.
   /// @param temperature in tempUnit
   /// @param humidity in percentage
   /// @param temptUnit in °C or °F. default=TempUnit_Celsius
-  /// @return Calculated heatindex as integer in TempUnit
-  int HeatIndex(
+  /// @return Calculated heatindex as float in TempUnit
+  float HeatIndex(
     float temperature,		
     float humidity,              
     TempUnit tempUnit = TempUnit_Celsius);
